tests/submission/PhilipTaylor/canvas/type.replace.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: type.replace</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.type.html">type</a>.replace</h1>
<p class="desc">HTMLCanvasElement methods can be replaced, and the replacement methods used by canvases</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.canvas.type">canvas.type</a>
</li></ul>
</div>
<p class="notes">Defined in "Web IDL" (draft)
</p><p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("HTMLCanvasElement methods can be replaced, and the replacement methods used by canvases");
_addTest(function(canvas, ctx) {

window.HTMLCanvasElement.prototype.getContext = function (name) { return 0; };
_assertSame(canvas.getContext('2d'), 0, "canvas.getContext('2d')", "0");


});
</script>

<div id="log"></div>
</body></html>
